agents-server/pyproject.toml
[project]
name = "agents-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26.4",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "huggingface-hub>=0.28.0",
    "jinja2>=3.1.4",
    "rich>=13.9.4",
    "python-dotenv",
    "duckduckgo-search>=6.3.7",
    "arize-phoenix",
    "opentelemetry-sdk", 
    "opentelemetry-exporter-otlp",
    "openinference-instrumentation-smolagents>=0.1.4",
    "openai>=1.58.1",
]
